fix: Refatorações importantes: Segurança de secrects | Alterações no backend | alterações no frontend | nova versão do banco de dados

+++ apiPenduraAi/src/main/java/com/ufc/apiPenduraAi/dtos/user/ReturnLoginDTO.java
@@ -1,4 +1,10 @@
 package com.ufc.apiPenduraAi.dtos.user;
 
-public record ReturnLoginDTO(String token, int id, String email, String nome, String role) {
+public record ReturnLoginDTO(
+        String token,
+        Long id,
+        String email,
+        String nome,
+        String role
+) {
 }
